Move stuff from TropicalFishUtils to TropicalFishEntity
